Fix password piping to grcryptfs process
The printf command which pipes the password to gocryptfs, has the password variable as format argument. This lets passwords with a %-character followed by the appropriate character be interpreted as format specification. For example: foo%d will be written as foo0 by printf. 
Inserting %s as format string and the password variable as its argument will ensure the correct printf output.